backend/scripts/requirements.txt
boto3==1.24.84
botocore==1.27.84
moviepy==1.0.3
requests==2.28.1
webvtt_py==0.4.6
